separate base protocol, debug viewer
