+++ cryodrgn/constants.py
@@ -34,7 +34,7 @@ V1_1_0 = "1.1.0"
 V2_0_0 = "2.0.0"
 
 VERSIONS = [V1_0_0, V1_1_0, V2_0_0]
-CRYODRGN_DEFAULT_VER_NUM = V1_1_0
+CRYODRGN_DEFAULT_VER_NUM = V2_0_0
 
 DEFAULT_ENV_NAME = getCryoDrgnEnvName(CRYODRGN_DEFAULT_VER_NUM)
 DEFAULT_ACTIVATION_CMD = 'conda activate ' + DEFAULT_ENV_NAME
